aider: Refactored `TripletDataset` to an iterable dataset.
# Aider chat conversation:

USER: The `TripletDataset` should be refactored to an iterable dataset.

ASSISTANT: To refactor the `TripletDataset` class to an iterable dataset, we will need to modify the `script/representation.py` file. The `torch.utils.data.IterableDataset` class will be used as a base class for `TripletDataset` instead of `torch.utils.data.Dataset`. We will also implement the `__iter__` method to yield data samples.

Here are the *SEARCH/REPLACE* blocks for the changes:

script/representation.py
```python
<<<<<<< SEARCH
class TripletDataset(torch.utils.data.Dataset):
    # TODO Switch to iterable dataset?
    def __init__(self, input_features, config_features):
        self.num_inputs = input_features.shape[0]
        self.num_configs = config_features.shape[0]

        self.input_features = input_features
        self.config_features = config_features

    def __len__(self):
        # TODO Calculate max. length
        return 12345

    def __getitem__(self, idx):
        return None, None
=======
class TripletDataset(torch.utils.data.IterableDataset):
    def __init__(self, input_features, config_features, pairs):
        super(TripletDataset, self).__init__()
        self.input_features = input_features
        self.config_features = config_features
        self.pairs = pairs

    def __iter__(self):
        for t, a, p, n in self.pairs:
            anchor = self.input_features[a] if t[0] == 'i' else self.config_features[a]
            positive = self.input_features[p] if t[1] == 'i' else self.config_features[p]
            negative = self.input_features[n] if t[2] == 'i' else self.config_features[n]
            yield anchor, positive, negative
>>>>>>> REPLACE
```

Please note that the `__len__` method is not typically implemented for `IterableDataset` as the length may not be known or relevant. The `__iter__` method will yield one triplet at a time from the `pairs` data structure. The `pairs` variable is assumed to be a list of tuples where each tuple contains the type of the sample ('i' for input or 'c' for configuration) and the indices of the anchor, positive, and negative samples. The `input_features` and `config_features` are assumed to be indexable collections that can return the feature vector for a given index.